Fix types

+++ src/server.c
@@ -5909,7 +5909,7 @@ sds genRedisInfoString(dict *section_dict, int all_sections, int everything) {
             "keyspace_misses:%lld\r\n"
             "pubsub_channels:%ld\r\n"
             "pubsub_patterns:%lu\r\n"
-            "pubsub_prefixes:%lu\r\n"
+            "pubsub_prefixes:%llu\r\n"
             "pubsubshard_channels:%lu\r\n"
             "latest_fork_usec:%lld\r\n"
             "total_forks:%lld\r\n"
@@ -5965,7 +5965,7 @@ sds genRedisInfoString(dict *section_dict, int all_sections, int everything) {
             server.stat_keyspace_misses,
             dictSize(server.pubsub_channels),
             dictSize(server.pubsub_patterns),
-            raxSize(server.pubsub_prefixes),
+            (unsigned long long) raxSize(server.pubsub_prefixes),
             dictSize(server.pubsubshard_channels),
             server.stat_fork_time,
             server.stat_total_forks,
